Extract RadarMeta to separate model
